refactor: Refactor event routes
